requirements.txt
pandas==2.1.1
streamlit==1.24.1
plotly-express==5.9.0
altair<5